To: Maurizio Boriani <baux@member.fsf.org> Cc: guile-user@gnu.org, guile-dbi-users@gna.org guile-dbi-devel@gna.org Subject: [PATCH 6/12] Guile-DBI: Fix crash, avoid recursive free Date: 15 Sept 2008
There is a particularly nasty crash scenario invloving the
garbage collector. The problem is that the dbi free routine
calls the dbi close routine, which tries to alloc new SCM
nodes. This is bad, because one should not alloc while
the garbage collector is running. So add a flag, indicating
that a free is in progress.

The DBD layers will have to honour this flag.

(also fixes one more mem leak).

Signed-off-by: Linas Vepstas <linasvepstas@gmail.com>

---
 include/guile-dbi/guile-dbi.h |    1 +
 src/guile-dbi.c               |   14 ++++++++++++--
 2 files changed, 13 insertions(+), 2 deletions(-)