To: Maurizio Boriani <baux@member.fsf.org> Cc: guile-user@gnu.org, guile-dbi-users@gna.org guile-dbi-devel@gna.org Subject: [PATCH 12/12] Guile-DBD-postgres: Crash in GC Date: 15 Sept 2008
Fix a serious bug which results in a crash in the GC. The problem
is that the DBI free routine will call the close routine, which will
call this routine, which tries to alloc new SCM cells. But one must
not alloc while the garbage collector is runing. Soo.. A previous
patch added a flag to indicate that we're in the GC. If this flag
is set, then avoid doing the cell allocs.

Signed-off-by: Linas Vepstas <linasvepstas@gmail.com>

---
 src/guile-dbd-postgresql.c |   14 ++++++++++----
 1 file changed, 10 insertions(+), 4 deletions(-)