fix-dlsym-dlerror-interaction-for-solaris.patch
This patch fixes a bug which causes any prior dlsym errors to propagate
into guile-dbi.  The fix is to call dlerror() immediately before
calling dlsym(), in order to clear the state.  The bug manifests on
Solaris 10, and may also be a problem on Linux.

From Andrew Gaylard <ag@computer.org>